Keep plugin configs in common engine configuration
